renamed class policy, add 2 class for 2 policies
